Small restructurization and stop application counter
