+++ server/cmd/root.go
@@ -42,6 +42,7 @@ func Execute(tag string) {
 	metricsRelay := metrics.InitRelay(clusterID, nil)
 	if metricsRelay != nil {
 		metrics.InitReplay(true)
+		defer metrics.StopRunningCounter()
 		interval := metricsRelay.Timeout
 		trigger := metrics.TickerTrigger{
 			Ticker: time.NewTicker(interval),
